Fixed Bugs

+++ FVCalc.java
@@ -1,8 +1,8 @@
 // Computes the future value of a saving investment.
 public class FVCalc {
 	public static void main(String[] args){
-		int deposit = Integer.parseInt(args[0]);
-		double percent = (double) Integer.parseInt(args[1]) / 100.0;
+		double deposit = Double.parseDouble(args[0]);
+		double percent = (double) Double.parseDouble(args[1]) / 100.0;
 		int years = Integer.parseInt(args[2]);
 		double yield = deposit;
